Changed: Now using Premake5 to generate project files

+++ BagualEngine/include/Definitions.h
@@ -3,7 +3,7 @@
 namespace bgl
 {
 	constexpr float kInfinity = std::numeric_limits<float>::max();
-	constexpr float kEpsilon = 1e-8;
+	constexpr float kEpsilon = static_cast<float>(1e-8);
 
 	typedef unsigned int uint;
 	typedef unsigned short ushort;
